refactor(size): keep O-D/O-F additions inside the module ratchet by extraction, not exemption
The three touched modules sat exactly at their gates, so the new code
moved to coherent leaves instead of buying debt:

- usage_accounting (1600-line gate): the settlement-time density observer
  moved beside its store as capability_evidence.observe_token_density;
  the historical _observe_token_density seam delegates, so both
  settlement call sites and monkeypatch targets keep working.
- tools/subagent_integration (1600-line gate): the one patch-verdict
  writer (artifact + custody row, shared by both patch pipelines) is now
  tools/patch_verdict.write_patch_verdict with the historical private
  name re-exported.
- llm.py (200K-byte gate): the bounded message-chars measure moved
  beside its proxy constant as context_budget.estimate_message_chars,
  aliased under the historical private name.

No behavior change; suites for all three surfaces stay green.
