+++ ouroboros/tools/subagent_integration.py
@@ -24,7 +24,7 @@ import tempfile
 from typing import Any, Dict, List, Tuple, Union
 
 from ouroboros.tools.registry import ToolContext, ToolEntry
-from ouroboros.artifacts import task_artifact_dir_path, task_id_for_artifacts
+from ouroboros.artifacts import task_artifact_dir_path
 from ouroboros.task_results import load_task_result
 from ouroboros.review_state import invalidate_advisory_after_mutation
 from ouroboros.runtime_mode_policy import (
@@ -39,7 +39,6 @@ from ouroboros.headless import (
     ARTIFACT_STATUS_READY_NO_CHANGES,
     ARTIFACT_STATUS_READY_WITH_CHANGES,
 )
-from ouroboros.utils import atomic_write_json, utc_now_iso
 
 log = logging.getLogger(__name__)
 
@@ -128,70 +127,9 @@ def _sha256_file(path: pathlib.Path) -> str:
     return hasher.hexdigest()
 
 
-def _write_verdict(
-    ctx: ToolContext,
-    child_task_id: str,
-    *,
-    outcome: str,
-    reason: str,
-    files: List[str],
-    manifest: Dict[str, Any],
-    applied: bool,
-    conflicts: List[str],
-    protected: List[str],
-    target: str = "",
-) -> str:
-    parent_task_id = task_id_for_artifacts(ctx)
-    art_dir = task_artifact_dir_path(getattr(ctx, "drive_root", "."), parent_task_id, create=True)
-    verdict = {
-        "schema_version": 1,
-        "created_at": utc_now_iso(),
-        "tool": "integrate_subagent_patch",
-        "parent_task_id": parent_task_id,
-        "child_task_id": child_task_id,
-        "outcome": outcome,
-        "applied": bool(applied),
-        "reason": str(reason or ""),
-        "target_root": str(target or ""),
-        "files": list(files or []),
-        "protected_matches": list(protected or []),
-        "conflicts": list(conflicts or []),
-        "patch_sha256": str((manifest or {}).get("sha256") or ""),
-        "diffstat": str((manifest or {}).get("diffstat") or ""),
-    }
-    path = art_dir / f"subagent_patch_verdict_{child_task_id}.json"
-    artifact_write_failed = False
-    try:
-        atomic_write_json(path, verdict, trailing_newline=True)
-    except Exception:
-        artifact_write_failed = True
-    # D-trace: the apply/reject decision also lands as a typed row in the
-    # custody event log, so the acceptance packet builds the disposition
-    # section from ONE replayable store for both patch pipelines (the
-    # delegated pipeline's run-keyed PATCH_DISPOSED row rides beside it; the
-    # scanner splits by `tool`, so nothing double-counts). A failed artifact
-    # write is disclosed on the row rather than dying as a silent "".
-    try:
-        from ouroboros import delegate_custody as custody
-
-        custody.emit(getattr(ctx, "drive_root", "."), "delegate_run_patch_verdict", {
-            "run_id": "",
-            "task_id": parent_task_id,
-            "child_task_id": child_task_id,
-            # The delegated pipeline mints its verdict subject as run_<rid>
-            # (every _integrate_delegated_patch call site); classifying at
-            # the ONE writer keeps that convention local instead of leaking
-            # prefix-matching into readers.
-            "pipeline": "delegated" if child_task_id.startswith("run_") else "subagent",
-            "disposition": outcome,
-            "applied": bool(applied),
-            "reason": str(reason or "")[:600],
-            "patch_sha256": str(verdict.get("patch_sha256") or ""),
-            "verdict_artifact_write_failed": artifact_write_failed,
-        })
-    except Exception:
-        log.debug("subagent patch verdict custody row failed", exc_info=True)
-    return "" if artifact_write_failed else str(path)
+# Verdict writing lives in tools/patch_verdict.py (module-size ceiling);
+# the historical private name stays importable for callers and tests.
+from ouroboros.tools.patch_verdict import write_patch_verdict as _write_verdict  # noqa: E402
 
 
 def _child_write_root(child_result: Dict[str, Any]) -> str:
